Improve error message, with an arrow pointing to the 'failing' marble.
- Extract marble classes into seperate files.
