+++ src/Divverence.MarbleTesting.Akka/Properties/AssemblyInfo.cs
@@ -1,4 +1,7 @@
 ﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+
 [assembly: AssemblyTitle("Divverence.MarbleTesting.Akka")]
 [assembly: AssemblyDescription("Divverence.MarbleTesting Library for Akka.Net")]
+[assembly: InternalsVisibleTo("Divverence.MarbelTesting.Tests")]
