+++ src/Divverence.MarbleTesting/Properties/AssemblyInfo.cs
@@ -1,4 +1,6 @@
 ﻿using System.Reflection;
+using System.Runtime.CompilerServices;
+
 [assembly: AssemblyTitle("Divverence.MarbleTesting")]
 [assembly: AssemblyDescription("Divverence.MarbleTesting Core Library")]
-
+[assembly: InternalsVisibleTo("Divverence.MarbleTesting.Tests")]
