+++ tst/Divverence.MarbleTesting.Tests/Divverence.MarbleTesting.Tests.csproj
@@ -42,6 +42,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="MarbleParserTest.cs" />
+    <Compile Include="ErrorMessageHelperTest.cs" />
     <Compile Include="MultiCharMarbleParserTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
